Resolve Nokogiri deprecation warning
This fixes the following deprecation warning:

```
worldpay_cnp-0.1.0/lib/worldpay_cnp/xml/nokogiri.rb:36: warning: Passing a Node as the second parameter to Node.new is deprecated. Please pass a Document instead, or prefer an alternative constructor like Node#add_child. This will become an error in a future release of Nokogiri.
```

We are now passing the document along with the parent element. The serializer will handle the case of setting the parent to the document, when the parent is nil due to the recursive nature of the method calls.

This deprecation was included in release 1.13.0 https://github.com/sparklemotion/nokogiri/releases/tag/v1.13.0

For reference: https://github.com/sparklemotion/nokogiri/issues/975 and https://github.com/sparklemotion/nokogiri/pull/2401